renamed some components
