+++ PhoneBookApp/PhoneBook/UI-Angular/AngularPhonebook/src/app/app-routing.module.ts
@@ -2,7 +2,7 @@ import { Component, NgModule } from '@angular/core';
 import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
 import { AuthGuard } from './Auth/guards/auth-guard.service';
 import { LoginComponent } from './Auth/login/login.component';
-import { SignalRAuthComponent } from './Auth/signal-r-auth/signal-r-auth.component';
+import { SignalRAuthComponent } from './signal-r/signal-r-auth.component';
 import { PhonebookComponent } from './phonebook/phonebook.component';
 import { ErrorPageComponent } from './shared/error-page/error-page.component';
 
@@ -11,7 +11,7 @@ const routes: Routes = [
   { path: 'login', component: LoginComponent },
   { path: 'phonebook', component:PhonebookComponent, canActivate: [AuthGuard] },
 
-  { path: 'chat', loadChildren: () => import('./Auth/signal-r-auth/signal-r/signal-r.module').then(m => m.SignalRModule),
+  { path: 'chat', loadChildren: () => import('./signal-r/signal-r/signal-r.module').then(m => m.SignalRModule),
     canActivate: [AuthGuard] },
   { path: '**', component: ErrorPageComponent } // must always be last.
 ]
