+++ PhoneBookApp/PhoneBook/UI-Angular/AngularPhonebook/src/app/signal-r/signalr-auth.service.ts
